Separando la ubicación de las funciones relacionadas al path y a los links
