+++ cli.js
@@ -1,6 +1,6 @@
 const { mdLinks } = require('./index.js');
 
-mdLinks('./carpeta de prueba')
+mdLinks('./carpeta de prueba/prueba/README-ghibli-data-lovers.md')
   .then(() => {})
   .catch((error) => {
     console.log(error);
